backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart==0.0.9
aiosqlite==0.20.0
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
APScheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
twilio==9.1.0
google-generativeai==0.5.4
